moving HoopoeProfilerBridge to its own source set and renaming
