+++ hoopoe-integration-testing/build.gradle.kts
@@ -76,7 +76,6 @@ tasks {
         }
     }
 
-
     "classes" {
         dependsOn("buildTestAgent", "buildHoopoeExtension")
     }
